finish delete method of guestbook
