+++ .gitignore
@@ -4,7 +4,7 @@
 /cypress/videos/
 .env
 /data/blogDrafts/
-docker-compose.yaml
+/docker/docker-compose.yaml
 
 # dependencies
 /node_modules
