+++ styles/global.css
@@ -13,7 +13,7 @@ This essentially allows us to make our own tailwind-like classNames. */
 }
 
 .my-h1 {
-  @apply font-bold text-3xl md:text-5xl tracking-tight mb-8 text-black dark:text-white;
+  @apply font-bold text-5xl tracking-tight mb-8 text-black dark:text-white;
 }
 .my-h2 {
   @apply font-bold text-3xl tracking-tight mb-8 mt-8 text-black dark:text-white;
